+++ core/test/test_vpn_manager_fsm_recovery.cpp
@@ -106,6 +106,8 @@ struct ConnectingVpnManagerTest : MockedTest {
         case VPN_EVENT_CONNECT_REQUEST:
         case VPN_EVENT_ENDPOINT_CONNECTION_STATS:
         case VPN_EVENT_DNS_UPSTREAM_UNAVAILABLE:
+        case VPN_EVENT_TUNNEL_CONNECTION_STATS:
+        case VPN_EVENT_TUNNEL_CONNECTION_CLOSED:
             break;
         case VPN_EVENT_STATE_CHANGED: {
             auto *event = (VpnStateChangedEvent *) data;